requirements.txt
# paddlepaddle-gpu
paddlepaddle==2.5.2
paddleocr==2.8.1
pyheif==0.7.1
opencc==1.1.6